Convert to LIBGDX!! Bugs galore but mostly working. Performance many times better

+++ src/main/kotlin/screen/elements/GUIOutline.kt
@@ -15,6 +15,6 @@ class GUIOutline(parent: RootGUIElement,
     }
 
     override fun render() {
-        Renderer.renderEmptyRectangle(xPixel, yPixel, widthPixels, heightPixels, borderThickness = borderThickness)
+        Renderer.renderEmptyRectangle(xPixel, yPixel, widthPixels, heightPixels, borderThickness = borderThickness.toFloat())
     }
 }
